+++ tests/test_static_treecode.py
@@ -120,7 +120,9 @@ def brute_force_reference(corr, shear, dens, w_s, w_d):
             else:
                 index, ra, dec, _ = parents_and_centroids(pix, nside_b)
             ids = np.arange(ra.size, dtype=np.int64)
-            I, J, bins, c1, s1, c2, s2 = _compute_pairs_impl(ids, ra, dec, edges[b0 : b1 + 1])
+            I, J, bins, c1, s1, c2, s2, _ = _compute_pairs_impl(
+                ids, ra, dec, edges[b0 : b1 + 1]
+            )
             members = [pix[index == c] for c in range(ra.size)]
             for a, b, bb, e1, e2 in zip(I, J, bins + b0, c1 + 1j * s1, c2 + 1j * s2):
                 pa, pb = members[a], members[b]